server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dill==0.3.7
fastapi==0.103.0
fastapi-cache2==0.2.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
pendulum==2.1.2
platformdirs==3.10.0
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytzdata==2020.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
uvicorn==0.23.2
wrapt==1.15.0
yarl==1.9.2
